refactor: refactor server, extract storage and handler
